+++ shuup/admin/modules/media/static_src/media/browser/BrowserView.js
@@ -100,7 +100,18 @@ export function controller(config={}) {
     };
     ctrl.reloadFolderTree = function() {
         remote.get({"action": "folders"}).then(function(response) {
-            ctrl.rootFolder(response.rootFolder);
+            if (response.rootFolder.id === 0 && !response.rootFolder.canSeeRoot) {
+                // Hide the root folder if the user cannot access it.
+                // This is only a visual thing to avoid showing the empty folder.
+                // The view handles the actual permission checking and clears the whole file
+                // QueyrSet of the root folder if the user is not allowed to access it.
+                const currentFolder = response.rootFolder.children[0];
+                ctrl.rootFolder(currentFolder || {});
+                ctrl.currentFolderId(currentFolder ? currentFolder.id : null);
+                ctrl.reloadFolderContents();
+            } else {
+                ctrl.rootFolder(response.rootFolder);
+            }
             ctrl._refreshCurrentFolderPath();
         });
     };